requirements.txt
Django>=4.0
djangorestframework
python-dotenv
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-crispy-forms
crispy-bootstrap4
celery
redis
pdfminer.six
spacy
scikit-learn